pyproject.toml
[tool.poetry]
name = "MediaManager"
version = "0.1.0"
description = "Media Management Tools"
authors = ["Michael Genson <71845777+michael-genson@users.noreply.github.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.22.0"
fastapi = "^0.97.0"
httpx = "^0.24.1"
pydantic = "^1.10.9"
apprise = "^1.4.0"
pyhumps = "^3.8.0"
jinja2 = "^3.1.2"
rocketry = "^2.5.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.272"
mypy = "^1.3.0"
pre-commit = "^3.3.3"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120

[tool.ruff]
extend-exclude = ["deploy/*", "tests/*"]
line-length = 120
target-version = "py311"
